requirements.txt
# Core Framework
python-dotenv==1.0.0
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Database - Use prebuilt wheels only
pymongo==4.10.1

# Data Processing - Python 3.13 compatible prebuilt wheels
pandas==2.2.3
numpy==2.1.2

# Technical Analysis
ta==0.11.0

# ML (if needed)
scikit-learn==1.5.2

# Utilities
python-dateutil==2.9.0
pytz==2025.2

# Remove: pydantic, motor, joblib (not needed for basic ML engine)# Add these
tensorflow==2.17.0  # For LSTM (Keras)
joblib==1.4.2  # For saving RF models